split filtering and labelling modules, deleted AlleleSelector, wrote view class
